[test] Move unit tests for fake RSA keys to a separate TU
